Refactor: ProposalDetails components NNS agnostic
